+++ drivers/input/keyboard.h
@@ -1,0 +1,9 @@
+// drivers/input/keyboard.h
+#ifndef TOYIX_DRIVERS_INPUT_KEYBOARD_H
+#define TOYIX_DRIVERS_INPUT_KEYBOARD_H
+
+
+void keyboard_init(void);
+
+
+#endif